refactor(events): split the gateway dispatcher by responsibility
accord_event_handler.dart was the only first-party file over 1,000 lines. It
was already sectioned by domain, and the sections shared nothing but the
enclosing scope, so the cut follows the seams that were already there:

- accord_ready_sync.dart takes the READY bootstrap helpers, which were
  already top-level functions that only need a Ref. This is the REST
  hydration done once on connect, not event handling.
- accord_message_events.dart takes the message domain (incoming messages,
  message cache, read state, reactions, typing). bindMessageEvents appends
  to the same subscription list at the same point in the sequence, so
  registration order and the single disposer are unchanged. The four
  concerns inside the message.create listener stay in one closure in order.

The handler keeps connection lifecycle, READY, presence, users, voice,
spaces, channels, members and roles, and drops from 1087 to 518 lines.

Also here: member-leave now reuses the roleSpaceId helper instead of
repeating its expression, and the two READY list loads use the existing
listOrLog extension. No gateway stream gained or lost a subscriber; a new
test drives one message.create end to end to guard the split, which nothing
covered before.
